requirements.txt
flask==3.0.0
flask-sqlalchemy==3.1.1
flask-migrate==4.0.5
flask-wtf==1.2.1
flask-login==0.6.3
python-dotenv==1.0.0
pymysql==1.1.0
passlib==1.7.4
